V8PP: convert.hpp: automatically heap-copy and wrap temporary objects
When a non-pointer, non-shared-pointer object needs to be converted to V8
and has not been previously registered, assume it is copyable. Create and
register a copy on the heap that can be passed to V8.
This fixes functions that return temporary objects, which previously
threw "failed to wrap C++ object".

Signed-off-by: Tim Janik <timj@gnu.org>